add jwt refresh token

+++ code/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 
 from models.item import ItemModel
 
@@ -41,6 +41,9 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'},401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -66,6 +69,14 @@ class Item(Resource):
 
 
 class ItemList(Resource):
-
+    @jwt_required(optional=True)
     def get(self):
-        return {"items": [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items':items}, 200
+
+        return {
+            "items": [item['name'] for item in items],
+            'message':'more data available if you log in!'
+        }
